# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Shopify Conectos API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/category": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Listar categorías planas",
                "operationId": "84e00c3f53533b76d7d1c4cadd28496d",
                "parameters": [
                    {
                        "name": "provider",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "level",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "parent_provider_category_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "active",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of categories",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "provider": {
                                                "type": "string",
                                                "example": "syscom"
                                            },
                                            "provider_category_id": {
                                                "type": "string",
                                                "example": "12"
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Cámaras IP nivel 12"
                                            },
                                            "level": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "active": {
                                                "type": "boolean",
                                                "example": true
                                            },
                                            "shopify_type": {
                                                "type": "string",
                                                "example": "collection"
                                            },
                                            "shopify_id": {
                                                "type": "string",
                                                "example": "gid://shopify/Collection/488809431263"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "id": {
                                                "type": "integer",
                                                "example": 15
                                            },
                                            "parent_provider_category_id": {
                                                "type": "string",
                                                "example": "12",
                                                "nullable": true
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "Crear categorías masivamente",
                "operationId": "cc38f307b015da73530146ccbe2edc00",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "required": [
                                        "provider",
                                        "provider_category_id",
                                        "name",
                                        "level"
                                    ],
                                    "properties": {
                                        "provider": {
                                            "type": "string",
                                            "example": "syscom"
                                        },
                                        "provider_category_id": {
                                            "type": "string",
                                            "example": "12"
                                        },
                                        "parent_id": {
                                            "type": "string",
                                            "example": null,
                                            "nullable": true
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Cámaras IP nivel 12"
                                        },
                                        "level": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object",
                                    "example": [
                                        {
                                            "provider": "syscom",
                                            "provider_category_id": "12",
                                            "name": "Cámaras IP nivel 12",
                                            "level": 1
                                        },
                                        {
                                            "provider": "syscom",
                                            "provider_category_id": "2",
                                            "parent_id": "12",
                                            "name": "Cámaras IP",
                                            "level": 2
                                        },
                                        {
                                            "provider": "syscom",
                                            "provider_category_id": "3",
                                            "parent_id": "2",
                                            "name": "Domo",
                                            "level": 3
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "ok": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "provider": {
                                                        "type": "string",
                                                        "example": "syscom"
                                                    },
                                                    "provider_category_id": {
                                                        "type": "string",
                                                        "example": "12"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Cámaras IP nivel 12"
                                                    },
                                                    "level": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "active": {
                                                        "type": "boolean",
                                                        "example": true
                                                    },
                                                    "shopify_type": {
                                                        "type": "string",
                                                        "example": "collection"
                                                    },
                                                    "shopify_id": {
                                                        "type": "string",
                                                        "example": "gid://shopify/Collection/488809431263"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 15
                                                    },
                                                    "parent_provider_category_id": {
                                                        "type": "string",
                                                        "example": "12",
                                                        "nullable": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/category/tree": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Obtener árbol de categorías",
                "operationId": "f07f72142f9ee915b4cfa426876cede2",
                "parameters": [
                    {
                        "name": "provider",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "active",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category Tree",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "ok": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "provider": {
                                                        "type": "string",
                                                        "example": "syscom"
                                                    },
                                                    "provider_category_id": {
                                                        "type": "string",
                                                        "example": "12"
                                                    },
                                                    "parent_provider_category_id": {
                                                        "type": "string",
                                                        "example": null,
                                                        "nullable": true
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Cámaras IP"
                                                    },
                                                    "level": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "active": {
                                                        "type": "boolean",
                                                        "example": true
                                                    },
                                                    "shopify_type": {
                                                        "type": "string",
                                                        "example": "collection"
                                                    },
                                                    "shopify_id": {
                                                        "type": "string",
                                                        "example": "gid://shopify/Collection/123"
                                                    },
                                                    "children": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your bearer token in the format **Bearer &lt;token&gt;**",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}